+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Items/LithiumBatteries.cs
@@ -1,5 +1,4 @@
-﻿using MoreShipUpgrades.Configuration;
-using MoreShipUpgrades.Configuration.Upgrades.Interfaces.TierUpgrades;
+﻿using MoreShipUpgrades.Configuration.Upgrades.Interfaces.TierUpgrades;
 using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc.Upgrades;
 using MoreShipUpgrades.Misc.Util;