+++ ChangeLog.md
@@ -1,10 +1,12 @@
 
 <details>
-<summary> 3.12.11 - 2026-??-?? </summary>
+<summary> 3.12.11 - 2026-02-13 </summary>
 
 - Added Scrap Keeper configuration to only apply its effect on items that were previously stored in the ship and ignore items stored in the same day.
 - Implemented "Item Duplicator" which adds a chance of duplicating purchased items from the Company Store when dropped from the item drop pod.
 - Implemented "Baby Pacifier" which reduces the growth rate of maneater when they are crying.
+- Implemented "Smarter Lockpick" which increases the efficiency of the lockpicker item, decreasing the amount of time required to open the locked door.
+- Possibly fixed Landing Thrusters causing slow landings after landing on the Company moon.
 
 </details>